wrong syntax

+++ cookbooks/nginx/recipes/default.rb
@@ -15,12 +15,12 @@ if ['app', 'app_master'].include?(node[:instance_role])
 
 	execute "remove default config" do
 		command "rm /etc/nginx/servers/sproutvideo.conf"
-		if { FileTest.exits?("/etc/nginx/servers/sproutvideo.conf") }
+		only_if { FileTest.exits?("/etc/nginx/servers/sproutvideo.conf") }
 	end
 
 	execute "remove default ssl config" do
 		command "rm /etc/nginx/servers/sproutvideo.ssl.conf"
-		if { FileTest.exists?("/etc/nginx/servers/sproutvideo.ssl.conf")}
+		only_if { FileTest.exists?("/etc/nginx/servers/sproutvideo.ssl.conf") } 
 	end
 
 	execute "restarting nginx" do
